* Applications/15.12:
  Stop crashing when clizip finds a folder called //
  Do not show a messagebox with empty error message
  Fix opening of archives whose mimetype cannot be detected-by-content
  GIT_SILENT Upgrade KDE Applications version to 15.12.0.
  Show correct uncompressed size of huge zipped files

+++ plugins/clizipplugin/cliplugin.cpp
@@ -161,7 +161,7 @@ bool CliPlugin::readListLine(const QString &line)
             //          ending with '/' is actually more reliable than 'd' bein in the attributes.
             e[IsDirectory] = rxMatch.captured(10).endsWith(QLatin1Char('/'));
 
-            e[Size] = rxMatch.captured(4).toInt();
+            e[Size] = rxMatch.captured(4);
             QString status = rxMatch.captured(5);
             if (status[0].isUpper()) {
                 e[IsPasswordProtected] = true;
